% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addPackageGithub}
\alias{addPackageGithub}
\title{Add DESCRIPTION information from package on github.}
\usage{
addPackageGithub(pdb = pkgAvail(), repo, username = NULL,
  branch = "master")
}
\arguments{
\item{pdb}{Package database, usually the result of \code{\link{pkgAvail}} or \code{\link{available.packages}}}

\item{repo}{Character vector. Name of repository on github, e.g. \code{"RevolutionAnalytics/RRT"}}

\item{username}{Optional character vector. Name of repository on github, e.g. \code{"RevolutionAnalytics/RRT"}}

\item{branch}{name of branch, defaults to \code{"master"}}
}
\description{
Downloads the DESCRIPTION file from a package on github, parses the fields and adds (or replaces) a row in the available package database.
}
\examples{
# Create package database
pdb <- pkgAvail(repos=c(CRAN="http://cran.revolutionanalytics.com"))

# Overwrite pdb with development version of miniCRAN at github
newpdb <- addPackageGithub(pdb, "andrie/miniCRAN")
newpdb["miniCRAN", ]

# Add package from github that's not currently on CRAN
newpdb <- addPackageGithub(pdb, "RevolutionAnalytics/rrt")


# Add package from github that's not currently on CRAN
newpdb <- addPackageGithub(repo="RevolutionAnalytics/RRT")

set.seed(1)
plot(makeDepGraph("RRT", availPkgs = newpdb, suggests=TRUE), 
     legendPosVertex=c(1, -1), legendPosEdge=c(1, -0.7))





}

