% Generated by roxygen2 (4.0.1): do not edit by hand
\name{basePkgs}
\alias{basePkgs}
\title{Returns names of base packages.}
\usage{
basePkgs()
}
\description{
Retrieves names of installed packages by calling \code{\link[utils]{installed.packages}} and returning only those packages where \code{Priority} equals "base".
}
\seealso{
pkgDeps
}

