% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{checkVersions}
\alias{checkVersions}
\title{Check for previous versions of packages in a miniCRAN repository.}
\usage{
checkVersions(pkgs = NULL, path = NULL, type = "source",
  Rversion = getRversion())
}
\arguments{
\item{pkgs}{Character vector of packages to be installed. If not provided,
             checks all files for multiple package versions.}

\item{path}{The local path to the directory where the miniCRAN repo resides.}

\item{type}{character, indicating the type of package to download and
 install. See \code{\link{install.packages}}.}

\item{Rversion}{numeric version of the R system for which to fetch packages.
See \code{\link{R_system_version}}.}
}
\value{
Returns filepaths to packages with multiple versions for removal.
}
\description{
Checks for previous versions, and returns the file paths for packages with
multiple versions. The admin can subsequently decide which version to keep.
}
\examples{
\inst\examples\example_checkVersions.R
}

