% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{cranJuly2014}
\alias{cranJuly2014}
\title{Stored version of available.packages()}
\format{matrix}
\usage{
cranJuly2014
}
\description{
Copy of the result of \code{\link[utils]{available.packages}} on July 1, 2014.
}
\keyword{datasets}

