% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P3M.R
\name{is.online}
\alias{is.online}
\title{Returns TRUE if the p3m URL can be accessed.}
\usage{
is.online(url = NULL, tryHttp = TRUE)
}
\arguments{
\item{url}{p3m url}

\item{tryHttp}{If TRUE, also attempts http URL, for compatibility with older versions of R}
}
\description{
Returns TRUE if the p3m URL can be accessed.
}
