% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPackages.R
\docType{methods}
\name{.listFiles}
\alias{.listFiles}
\title{List pre-built packages in a directory based on file extension}
\usage{
.listFiles(path, type)
}
\arguments{
\item{path}{Character sting specifying the directory containing packages to be added.}
}
\value{
Installs the packages and returns the new package index.
}
\description{
List pre-built packages in a directory based on file extension
}
\examples{
\dontrun{
 .listFiles('path/to/my/packages', type = "source")
}

}

