% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeLibrary}
\alias{makeLibrary}
\title{Deprecated function to download packages to local folder.}
\usage{
makeLibrary(pkgs, path, type = "source")
}
\arguments{
\item{pkgs}{Character vector of packages to download}

\item{path}{Destination download path. This path is the root folder of your
new repository.}

\item{type}{Possible values are (currently) "source", "mac.binary" and "win.binary": the binary types can be listed and downloaded but not installed on other platforms.  Passed to \code{\link{download.packages}}.}
}
\description{
Deprecated function to download packages to local folder.
}

