% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPackages.R
\docType{methods}
\name{addLocalPackage}
\alias{addLocalPackage}
\title{Add local packages to a miniCRAN repository.}
\usage{
addLocalPackage(pkgs, pkgPath, path, type = "source", Rversion = R.version,
  writePACKAGES = TRUE, deps = FALSE, quiet = FALSE, build = FALSE)
}
\arguments{
\item{pkgs}{Character vector of packages to download}

\item{pkgPath}{Character vector of directory location containing packages to be added.}

\item{path}{Destination download path. This path is the root folder of your new repository.}

\item{type}{Possible values are (currently) "source", "mac.binary" and "win.binary": the binary types can be listed and downloaded but not installed on other platforms.  Passed to \code{\link{download.packages}}.}

\item{Rversion}{numeric version of the R system for which to fetch packages. See \code{\link{R_system_version}}.}

\item{writePACKAGES}{If TRUE, calls \code{\link{write_PACKAGES}} to update the repository PACKAGES file.}

\item{deps}{logical indicating whether the package dependencies should be added (default \code{TRUE}).}

\item{quiet}{If TRUE, suppress status messages (if any), and the progress bar during download.}

\item{build}{Logical indicating whether packages should be build prior to adding.}
}
\value{
Installs the packages and returns the new package index.
}
\description{
Examine the contents of a directory specified by \code{pkgPath} for pre-built
packages matching the names specified by \code{pkgs}, and add these to the
miniCRAN repository.
}
\details{
To build a package from source and then add it, use \code{build = TRUE} (NOT YET IMPLEMENTED).
Note that package development libraries and the \code{devtools} package
must be installed on your system in order to build packages.
}
\note{
Currently, adding local packages does not check nor download their dependencies.
}
\examples{
\dontrun{
 addLocalPackage("myPackage", "path/to/my/prebuilt/package",
                 "path/to/my/miniCRAN/repo")
 addLocalPackage("myPackage", "path/to/my/package/sourcecode",
                 "path/to/my/miniCRAN/repo", build=TRUE)
}

}
\author{
Alex Chubaty
}

