% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPackages.R
\docType{methods}
\name{addPackageGitHub}
\alias{addPackageGitHub}
\title{Add package from GitHub}
\usage{
addPackageGitHub(repo, ...)
}
\arguments{
\item{repo}{Github repository name in the format \code{username/repo}.}

\item{...}{Additional arguments passed to \code{addLocalPackage()}.}
}
\description{
Add package from GitHub
}
\note{
Currently, only installing from master branch is supported.
}
\examples{
\dontrun{
  addPackageGitHub("andrie/miniCRAN", path = reposPath)
}
}
\author{
Marvin Steijaert

Alex Chubaty
}
