% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgDep.R
\name{basePkgs}
\alias{basePkgs}
\title{Returns names of base packages.}
\usage{
basePkgs()
}
\description{
Retrieves names of installed packages by calling \code{\link[utils]{installed.packages}} and returning only those packages where \code{Priority} equals "base".
}
\seealso{
\code{\link{pkgDep}}

Other dependency functions: \code{\link{makeDepGraph}},
  \code{\link{pkgDep}}, \code{\link{plot.pkgDepGraph}}
}

