% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletePackages.R
\name{deletePackage}
\alias{deletePackage}
\title{Delete packages from a miniCRAN repository.}
\usage{
deletePackage(pkgs = NULL, path = NULL, type = "source",
  Rversion = R.version, writePACKAGES = TRUE, deps = TRUE, ...)
}
\arguments{
\item{pkgs}{Character vector of packages to download}

\item{path}{Destination download path. This path is the root folder of your new repository.}

\item{type}{Possible values are (currently) "source", "mac.binary" and "win.binary": the binary types can be listed and downloaded but not installed on other platforms.  Passed to \code{\link[=download.packages]{download.packages()}}.}

\item{Rversion}{Version of R. Can be specified as a character string with the two digit R version, e.g. "3.1".  Defaults to \link{R.version}}

\item{writePACKAGES}{If TRUE, calls \code{\link[=write_PACKAGES]{write_PACKAGES()}} to update the repository PACKAGES file.}

\item{deps}{logical indicating whether the package reverse dependencies should be removed (default \code{TRUE}).}

\item{...}{arguments passed over to \code{tools::package_dependencies}: use \code{which}, \code{recursive}, \code{reverse}
to specify what should happen to dependencies of removed packages}
}
\description{
Delete packages from a miniCRAN repository.
}
\seealso{
Other update repo functions: \code{\link{addOldPackage}},
  \code{\link{addPackage}}, \code{\link{checkVersions}},
  \code{\link{makeRepo}}, \code{\link{updatePackages}}
}
\concept{update repo functions}
