% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCranDescription}
\alias{getCranDescription}
\title{Scrape DESCRIPTION from CRAN for each pkg.}
\usage{
getCranDescription(pkg, repos = getOption("repos"), type = "source", path,
  pkgs = pkgDep(pkg, repos = repos, type = type))
}
\arguments{
\item{pkg}{Character vector of packages.}

\item{repos}{URL(s) of the 'contrib' sections of the repositories. Passed to \code{\link{available.packages}}}

\item{type}{Possible values are (currently) "source", "mac.binary" and "win.binary": the binary types can be listed and downloaded but not installed on other platforms.  Passed to \code{\link{download.packages}}.}

\item{path}{Destination download path. This path is the root folder of your
new repository.}

\item{pkgs}{Character vector of packages to download}
}
\description{
Scrape DESCRIPTION from CRAN for each pkg.
}
\examples{
\dontrun{
getCranDescription(c("igraph", "ggplot2", "XML"), 
  repos=c(CRAN="http://cran.revolutionanalytics.com")
)
}
}

