% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPackages.R
\docType{methods}
\name{.listFiles}
\alias{.listFiles}
\title{List pre-built packages in a directory based on file extension}
\usage{
.listFiles(pkgs, path, type)
}
\arguments{
\item{pkgs}{Character vector of package names}

\item{path}{Character string specifying the directory containing packages to be added.}

\item{type}{Character indicating the package type (e.g., "source", "win.binary", etc.).}
}
\value{
Installs the packages and returns the new package index.
}
\description{
List pre-built packages in a directory based on file extension
}
\examples{
\dontrun{
 .listFiles('path/to/my/packages', type = "source")
}

}

