% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{miniCRAN-package}
\alias{miniCRAN}
\alias{miniCRAN-package}
\alias{minicran}
\title{Tools to create an internally consistent, mini version of CRAN with selected packages only.}
\description{
Tools to create an internally consistent, mini version of CRAN with selected packages only.
}
\details{
At the end of 2013, CRAN consisted of more than 5000 packages.  Many organisations need to maintain a private mirror of CRAN, but with only a subset of packages that are relevant to them.

\code{miniCRAN} makes this possible by recursively reading the dependency tree for a given set of packages, then downloading only this subset.

Important functions:

\itemize{
\item \code{\link{pkgAvail}}: Read local repository and determine available packages
\item \code{\link{pkgDep}}: Find (recursive) package dependencies
\item \code{\link{makeDepGraph}}: Create graph of selected package dependencies
\item \code{\link{makeRepo}} : Make repository
}

This subset will be internally consistent, i.e. the following functions will work as expected:

\itemize{
\item \code{\link{available.packages}}
\item \code{\link{install.packages}}
}

The main function is \code{\link{makeRepo}} - this will download all the required packages, with their dependencies, into the appropriate repository file structure, and then create the repository index (PACKAGES) file.

To get a recursive list of dependencies, and a plot, use \code{\link{pkgDeps}} and \code{\link{makeDepGraph}}.
}
\author{
Andrie de Vries \email{apdevries@gmail.com}
}
\seealso{
\code{\link{minicran}}
}
\keyword{package}

