% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plot.pkgDepGraph}
\alias{plot.pkgDepGraph}
\title{Plots a package dependency graph.}
\usage{
\method{plot}{pkgDepGraph}(x, pkgsToHighlight,
  main = "Package dependency graph", ...)
}
\arguments{
\item{x}{pkgDepGraph object}

\item{pkgsToHighlight}{Optional character vector with names of package to hightlight. If missing, defaults to packages used in original call to \code{\link{makeDepGraph}}}

\item{main}{Title of plot}

\item{...}{Ignored}
}
\description{
Plots a package dependency graph.
}
\seealso{
plot.pkgDepGraph
}

