% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{repoBinPath}
\alias{repoBinPath}
\title{Construct path to full binary location}
\usage{
repoBinPath(path, type, Rversion)
}
\arguments{
\item{path}{Destination download path. This path is the root folder of your
new repository.}

\item{type}{Possible values are (currently) "source", "mac.binary" and
"win.binary": the binary types can be listed and downloaded but not
installed on other platforms.  Passed to \code{\link[=download.packages]{download.packages()}}.}

\item{Rversion}{Version of R (only used if \code{type} is not \code{source}.) Defaults to \link{R.version}, but this can be specified as any of the following formats:
\itemize{
\item a character string with the two digit R version, e.g. "3.1"
\item a list with components \code{major} and \code{minor}
\item the result of \code{\link[=getRversion]{getRversion()}}
\item the result of \link{R.version}
}}
}
\description{
Construct path to full binary location
}
\keyword{Internal}
