% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{repoBinPath}
\alias{repoBinPath}
\title{Construct path to full binary location}
\usage{
repoBinPath(path, type, Rversion)
}
\arguments{
\item{path}{Destination download path. This path is the root folder of your new repository.}

\item{type}{character, indicating the type of package to download and install. See \code{\link{install.packages}}.}

\item{Rversion}{List with two named elements: `major` and `minor`. If not supplied, defaults to system version of R, using \code{\link{R.version}}.  Only used if \code{type} is not "source"}
}
\description{
Construct path to full binary location
}

