% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{repoBinPath}
\alias{repoBinPath}
\title{Construct path to full binary location}
\usage{
repoBinPath(path, type, Rversion)
}
\arguments{
\item{path}{Destination download path. This path is the root folder of your new repository.}

\item{type}{Possible values are (currently) "source", "mac.binary" and "win.binary": the binary types can be listed and downloaded but not installed on other platforms.  Passed to \code{\link[=download.packages]{download.packages()}}.}

\item{Rversion}{List with two named elements: \code{major} and \code{minor}. If not supplied, defaults to system version of R, using \link{R.version}.  Only used if \code{type} is not "source"}
}
\description{
Construct path to full binary location
}
