% Generated by roxygen2 (4.0.2): do not edit by hand
\name{repoPrefix}
\alias{repoPrefix}
\title{Get the path to the repo directory containing the package files.}
\usage{
repoPrefix(type, Rversion)
}
\value{
The filepath to the package files directory.
}
\description{
Get the path to the repo directory containing the package files.
}
\section{Repo folder structure}{

The folder structure of a repository
\itemize{
 \item{Root}
 \itemize{
   \item{src}
   \itemize{
     \item{contrib}
   }
 \item{bin}
 \itemize{
   \item{windows/contrib/}
   \item{macosx/contrib/}
   \item{macosx/mavericks/contrib}
   \item{macosx/leopard/contrib}
 }
 \item{PACKAGES}
 }
}
}

