% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{repoPrefix}
\alias{repoPrefix}
\title{Get the path to the repo directory containing the package files.}
\usage{
repoPrefix(type, Rversion)
}
\arguments{
\item{type}{character, indicating the type of package to download and install. See \code{\link{install.packages}}.}

\item{Rversion}{Version of R. Can be specified as a character string with the two digit R version, e.g. "3.1".  Defaults to \code{\link{R.version}}}
}
\value{
The filepath to the package files directory.
}
\description{
Get the path to the repo directory containing the package files.
}
\section{Repo folder structure}{

The folder structure of a repository
\itemize{
 \item{Root}
 \itemize{
   \item{src/contrib}
   \itemize{
     \item{PACKAGES}
   }
   \item{bin}
   \itemize{
     \item{windows/contrib/version}
     \itemize{
       \item{PACKAGES}
     }
     \item{macosx/contrib/version}
     \itemize{
       \item{PACKAGES}
     }
     \item{macosx/mavericks/contrib/version}
     \itemize{
       \item{PACKAGES}
     }
     \item{macosx/leopard/contrib/version}
     \itemize{
       \item{PACKAGES}
     }
   }
 }
}
}

