% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{twodigitRversion}
\alias{twodigitRversion}
\title{Get a two-digit version of the R version}
\usage{
twodigitRversion(Rversion = R.version)
}
\arguments{
\item{Rversion}{Version of R (only used if \code{type} is not \code{source}.) Defaults to \link{R.version}, but this can be specified as any of the following formats:
\itemize{
\item a character string with the two digit R version, e.g. "3.1"
\item a list with components \code{major} and \code{minor}
\item the result of \code{\link[=getRversion]{getRversion()}}
\item the result of \code{\link[=R.version]{R.version()}}
}}
}
\value{
A character string representing the two-digit R version.
}
\description{
Get a two-digit version of the R version
}
\keyword{Internal}
